fix tests structure
